backend/requirements.txt
fastapi
python-multipart
uvicorn
langchain
openai
supabase
PyPDF2
python-docx
python-dotenv
langchain-community